added claude tests
